requirements.txt
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
curlify==3.0.0
dynaconf==3.2.11
exceptiongroup==1.3.0
Faker==37.4.0
Flask==3.1.1
Flask-JWT-Extended==4.7.1
future-fstrings==1.2.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
networkx==3.4.2
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
PySocks==1.7.1
pytest==8.4.0
pytest-depends==1.0.1
pytz==2025.2
requests==2.32.3
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
allure-python-commons~=2.15.0
pip~=25.1.1
wheel~=0.41.2
typing_extensions~=4.14.1
setuptools~=68.2.0